pyproject.toml
[project]
name = "lesson-raspberry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.119.0",
    "flask>=3.1.2",
    "jinja2>=3.1.6",
    "pygame>=2.6.1",
    "pymongo>=4.15.3",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.38.0",
]
